+++ Android/SheetMusicAudioAndroid/src/main/kotlin/io/github/jiyimeta/sheetmusic/audio/model/ScoreElementID.kt
@@ -15,4 +15,5 @@ sealed class ScoreElementID {
     data class Slur(val arg0: SlurID) : ScoreElementID()
     data class Jump(val staff: StaffAddress, val measureIndex: Int, val index: Int) : ScoreElementID()
     data class Marker(val staff: StaffAddress, val measureIndex: Int, val index: Int) : ScoreElementID()
+    data class Glissando(val start: NoteID) : ScoreElementID()
 }