+++ Sources/SheetMusicLayout/Selection/LayoutElement+ElementID.swift
@@ -8,6 +8,7 @@ extension LayoutElement {
     public var elementID: ScoreElementID? {
         switch self {
         case let .tieArc(_, _, _, identity): return identity
+        case let .glissandoLine(_, _, _, _, start): return start.map { .glissando(start: $0) }
         case let .marker(_, _, _, identity), let .jump(_, _, identity): return identity
         case let .keySignature(_, _, _, _, _, identity), let .timeSignature(_, _, _, _, identity):
             return identity