+++ Sources/SheetMusicLayout/Selection/ScoreHitTester.swift
@@ -94,7 +94,7 @@ public struct ScoreHitTester: Sendable {
         case let .clef(anchor): return .clef(anchor)
         case let .graceNote(id): return .graceNote(id)
         case .dynamic, .fermata, .breath, .tempo, .spanner, .keySignature, .timeSignature, .barLine, .articulation,
-             .tie, .slur, .jump, .marker:
+             .tie, .slur, .jump, .marker, .glissando:
             return target.elementID.map(ScoreItemID.element)
         case .stem, .flag, .beam, .lyric, .staffText, .harmony, .rehearsalMark:
             return nil